+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Manager helpers
   - getDeployFees
   - getRunFees
+- runWithMessage function
 
 ## 0.1.0 Nov 20, 2020